Fix clang-tidy warning about implicit long-to-int cast
This is harmless as int and long have the same size on all contemporary
platforms and the flag constants all fit into 32-bit int, but still
results in a warning, so suppress it with an explicit cast.

+++ src/tipwin.cpp
@@ -36,7 +36,7 @@ TipWindow::TipWindow(wxWindow *parent,
                      const wxSize& size,
                      long style,
                      const wxString& name)
-  : wxPopupWindow(parent, style),
+  : wxPopupWindow(parent, static_cast<int>(style)),
     m_text(text),
     m_margin(3, 3)
 {
